Validation notice spacing
I added another line break to each validation message. I felt this gave
the user a better idea of the issues they have to fix.

Is there a way to bullet these validation messages?

Should we change the color for the text, "Please fix the following:" to
make it stand out more?