Popup background color and close button
I have a lighter gray for the background of the popups and for the close
buttons of the popups. The only problem is that now the close button for
the instruments is a lighter gray than the background is. Also, I
noticed that the color for the font on the notify checkbox was a darker
gray than the agree checkbox. I figured we would want them both to be
almost white. So, I changed it. If they both need to revert back to the
darker gray, please let me know and I'll make the changes.